test: add configure option --disable-persistent
Some TPM commands require a reset after running to leave the TPM in a
clean state. If the tests are run without setting up the simulator
anew for each test, one might want to disable these.
Thus, a configure option --disable-persistent is added for disabling
such tests. If this option is given, TPM2TOOLS_TEST_PERSISTENT is set
to false. If (and only if) TPM2TOOLS_TEST_PERSISTENT is false, the
tests in question will return SKIP (77).
By default, those tests are enabled.

Signed-off-by: Johannes Holland <johannes.holland@infineon.com>